+++ pkg/core/src/tests/Adapter.t.sol
@@ -20,7 +20,7 @@ contract FakeAdapter is BaseAdapter {
         address _divider,
         address _target,
         address _oracle,
-        uint64 _ifee,
+        uint256 _ifee,
         address _stake,
         uint256 _stakeSize,
         uint256 _minm,
